pyproject.toml
[project]
name = "eks-hello-world"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aws-cdk-lambda-layer-kubectl-v31>=2.0.3",
    "aws-cdk-lib>=2.194.0",
    "boto3>=1.38.8",
    "constructs>=10.4.2",
    "keysentinel>=0.1.1",
]
